Event image upload and delete (#43)
* event image remove endpoint updated

* fixed the index to remove

+++ src/main/java/com/is4103/backend/controller/EventController.java
@@ -28,6 +28,7 @@ import com.is4103.backend.service.BusinessPartnerService;
 import com.is4103.backend.service.EventOrganiserService;
 import com.is4103.backend.service.EventService;
 import com.is4103.backend.service.SellerApplicationService;
+import com.is4103.backend.util.errors.EventImageNotFoundException;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -302,11 +303,24 @@ public class EventController {
         return allBps;
     }
 
+    // @PostMapping("/remove-pic")
+    // public ResponseEntity<String> removeEventPic(@RequestParam(name =
+    // "eid", defaultValue = "1") Long eid,
+    // @RequestParam(name = "imageIndex", defaultValue = "0") int imageIndex) {
+    // Event e = getEventById(eid);
+    // return eventService.removePicture(e, imageIndex);
+    // }
+
     @PostMapping("/remove-pic")
-    public ResponseEntity<String> getNewAppliationsFromEvent(@RequestParam(name = "eid", defaultValue = "1") Long eid,
-            @RequestParam(name = "imageIndex", defaultValue = "0") int imageIndex) {
+    public ResponseEntity<String> removeEventPic(@RequestParam(name = "eid", defaultValue = "1") Long eid,
+            @RequestParam(name = "imageUrl", defaultValue = "") String imageUrl) throws EventImageNotFoundException {
         Event e = getEventById(eid);
-        return eventService.removePicture(e, imageIndex);
+        int imageIndex = e.getImages().indexOf(imageUrl);
+        if (imageIndex < 0) {
+            throw new EventImageNotFoundException();
+        } else {
+            return eventService.removePicture(e, imageIndex);
+        }
     }
 
     @GetMapping("/categories")
